+++ packages/create-mechanic/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Updated canvas based templates to use the new `getCanvas` function
+
 ## 2.0.0-beta.9 - 2022-08-12
 
 ### Fixed